fix: useHook 추가 분리
