--- src/components/hooks/useAsync.js
@@ -1,22 +1,0 @@
-import { useState } from 'react';
-
-function useAsync(asyncFunction) {
-  const [pending, setPending] = useState(false);
-  const [error, setError] = useState(null);
-
-  const wrappedFunction = async (...args) => {
-    setPending(true);
-    setError(null);
-    try {
-      return await asyncFunction(...args);
-    } catch (error) {
-      setError(error);
-    } finally {
-      setPending(false);
-    }
-  };
-
-  return [pending, error, wrappedFunction];
-}
-
-export default useAsync;
